requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kaleido==0.2.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
pillow==11.0.0
plotly==5.24.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.24.0
tenacity==9.0.0
urllib3==2.2.3
Werkzeug==3.1.3
